+++ src/us/kbase/workspace/test/workspace/WorkspaceTest.java
@@ -6851,21 +6851,28 @@ public class WorkspaceTest extends WorkspaceTester {
 		assertOrdered(new ListObjectsParameters(Arrays.asList(wsi1, wsi2))
 				.withMaxObjectID(6L).withMinObjectID(1L), true);
 		
+		/* 19/2/13: This is really difficult to test on mongo 3+. There will always be a list
+		 * of workspaces in an $in clause, and so the query optimizer is now good enough to always
+		 * use the ws/obj/ver index to sort, at least for smaller data sets. It's still potentially
+		 * dangerous to add the sort, since that forces the optimizer to use the ws/obj/ver index
+		 * (which could return a huge number of results and really slow down the query).
+		 */
+		
 		//unsorted (at least with descending versions)
 		// type filter
-		assertOrdered(new ListObjectsParameters(user, SAFE_TYPE1), false);
+//		assertOrdered(new ListObjectsParameters(user, SAFE_TYPE1), false);
 		// after date filter
-		assertOrdered(new ListObjectsParameters(Arrays.asList(wsi1, wsi2))
-				.withAfter(Date.from(Instant.now().minusSeconds(100))), false);
+//		assertOrdered(new ListObjectsParameters(Arrays.asList(wsi1, wsi2))
+//				.withAfter(Date.from(Instant.now().minusSeconds(100))), false);
 		// before date filter
-		assertOrdered(new ListObjectsParameters(Arrays.asList(wsi1, wsi2))
-				.withBefore(Date.from(Instant.now())), false);
+//		assertOrdered(new ListObjectsParameters(Arrays.asList(wsi1, wsi2))
+//				.withBefore(Date.from(Instant.now())), false);
 		// user filter
-		assertOrdered(new ListObjectsParameters(Arrays.asList(wsi1, wsi2))
-				.withSavers(Arrays.asList(user)), false);
+//		assertOrdered(new ListObjectsParameters(Arrays.asList(wsi1, wsi2))
+//				.withSavers(Arrays.asList(user)), false);
 		// meta filter
-		assertOrdered(new ListObjectsParameters(Arrays.asList(wsi1, wsi2))
-				.withMetadata(new WorkspaceUserMetadata(meta)), false);
+//		assertOrdered(new ListObjectsParameters(Arrays.asList(wsi1, wsi2))
+//				.withMetadata(new WorkspaceUserMetadata(meta)), false);
 	}
 
 	private void assertOrdered(final ListObjectsParameters params, final boolean expectOrdered)